pyproject.toml
[project]
name = "gello-software"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "depthai>=3.2.1",
    "dm-control>=1.0.34",
    "numpy-quaternion>=2024.0.13",
    "omegaconf==2.3.0",
    "pillow>=12.0.0",
    "pin==3.7.0",
    "pure-python-adb>=0.3.0.dev0",
    "pygame>=2.6.1",
    "pyqt6>=6.10.1",
    "pyquaternion>=0.9.9",
    "pyrealsense2>=2.56.5.9235",
    "pyspacemouse>=1.1.5",
    "pyzmq>=27.1.0",
    "quaternion>=3.5.2.post4",
    "setuptools>=80.9.0",
    "termcolor>=3.2.0",
    "tyro>=1.0.1",
    "ur-rtde>=1.6.2",
    "xarm>=0.0.4",
    "xarm-python-sdk>=1.17.3",
    "zmq>=0.0.0",
]
